# openapi.yaml
openapi: 3.1.0
info:
  title: AGP Control Plane API
  version: 0.2.0
  description: >
    OpenAPI contract derived from the authoritative AGP control-plane specs.
    This document defines the network contract for orchestration clients and runtimes.
    V1 authentication uses bearer tokens only; /health is intentionally unauthenticated.
servers:
  - url: http://127.0.0.1:7860
...
security:
  - bearerAuth: []
paths:
...
  /runtimes:
    get:
      tags: [runtimes]
      summary: List runtimes
      operationId: listRuntimes
      deprecated: true
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/RuntimeStatus'
        - in: query
          name: health_status
          schema:
            $ref: '#/components/schemas/HealthStatus'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Runtime list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthenticated'
...
  /ops/runtimes:
    get:
      tags: [ops]
      summary: List managed runtimes
      operationId: opsListRuntimes
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/RuntimeStatus'
        - in: query
          name: health_status
          schema:
            $ref: '#/components/schemas/HealthStatus'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Runtime list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /ops/runtimes/{runtime_id}:
    get:
      tags: [ops]
      summary: Get runtime details
      operationId: opsGetRuntime
      parameters:
        - $ref: '#/components/parameters/RuntimeId'
      responses:
        '200':
          description: Runtime detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /ops/runtimes/{runtime_id}/drain:
    post:
      tags: [ops]
      summary: Drain a runtime
      operationId: opsDrainRuntime
      parameters:
        - $ref: '#/components/parameters/RuntimeId'
      responses:
        '200':
          description: Runtime draining
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /ops/runtimes/{runtime_id}/restart:
    post:
      tags: [ops]
      summary: Restart a runtime
      operationId: opsRestartRuntime
      parameters:
        - $ref: '#/components/parameters/RuntimeId'
      responses:
        '200':
          description: Runtime restarted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthenticated'
...
  /ops/alerts:
    get:
      tags: [ops]
      summary: Active alerts
      operationId: opsAlerts
      responses:
        '200':
          description: Alert list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /ops/audit:
    get:
      tags: [ops]
      summary: Audit log
      operationId: opsAudit
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Audit events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '401':
          $ref: '#/components/responses/Unauthenticated'
...
  /ops/alerts/dispatch:
    post:
      tags: [ops]
      summary: Dispatch alerts to configured webhook
      operationId: opsDispatchAlerts
      responses:
        '200':
          description: Alerts dispatched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /ops/jobs/{job_id}/trace:
    get:
      tags: [ops]
      summary: Full execution trace for a job
      operationId: opsJobTrace
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Job trace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /ops/logs/control-plane:
    get:
      tags: [ops]
      summary: Control plane logs
      operationId: opsControlPlaneLogs
      parameters:
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Log entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /ops/logs/runtimes/{runtime_id}:
    get:
      tags: [ops]
      summary: Runtime logs
      operationId: opsRuntimeLogs
      parameters:
        - $ref: '#/components/parameters/RuntimeId'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Log entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /ops/triage:
    get:
      tags: [ops]
      summary: Triage dashboard
      operationId: opsTriage
      responses:
        '200':
          description: Triage data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '401':
          $ref: '#/components/responses/Unauthenticated'
...
  /ops/upgrade-status:
    get:
      tags: [ops]
      summary: Persisted release and schema version state
      operationId: opsUpgradeStatus
      responses:
        '200':
          description: Upgrade status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpgradeStatusEnvelope'
        '401':
          $ref: '#/components/responses/Unauthenticated'
components:
  parameters:
...
    Cursor:
      in: query
      name: cursor
      schema:
        type: string
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 100
    JobId:
      in: path
      name: job_id
      required: true
      schema:
        type: string
...
    RuntimeId:
      name: runtime_id
      in: path
      required: true
      schema:
        type: string
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    Unauthenticated:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    ErrorCode:
      type: string
      enum:
        - invalid_request
        - not_found
        - conflict
        - stale_fencing_token
        - lease_expired
        - already_terminal
        - not_eligible
        - unauthenticated
        - forbidden
        - rate_limited
        - internal_error
...
    RuntimeStatus:
      type: string
      enum: [registering, idle, busy, degraded, offline, draining]
    HealthStatus:
      type: string
      enum: [healthy, degraded, unreachable, draining]
...
    ErrorEnvelope:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          const: false
        error:
          type: object
          required: [code, message, retryable]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
            retryable:
              type: boolean
    PageInfo:
      type: object
      required: [has_more]
      properties:
        next_cursor:
          type: string
        has_more:
          type: boolean
...
    RuntimeSummary:
      type: object
      required: [runtime_id, hostname, release_version, status, health_status, last_heartbeat_at]
      properties:
        runtime_id:
          type: string
        hostname:
          type: string
        release_version:
          type: string
        status:
          $ref: '#/components/schemas/RuntimeStatus'
        health_status:
          $ref: '#/components/schemas/HealthStatus'
        last_heartbeat_at:
          type: string
          format: date-time
        metadata_json:
          type: object
          additionalProperties: true
...
    RuntimeEnvelope:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/RuntimeSummary'
    RuntimeListEnvelope:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          const: true
        data:
          type: object
          required: [items, page]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/RuntimeSummary'
            page:
              $ref: '#/components/schemas/PageInfo'
    UpgradeStatus:
      type: object
      required: [release_version, schema_version, package_version]
      properties:
        release_version:
          type: string
        schema_version:
          type: string
        previous_release_version:
          type: string
        previous_schema_version:
          type: string
        package_version:
          type: string
        rollback_target_release_version:
          type: string
        rollback_target_schema_version:
          type: string
    UpgradeStatusEnvelope:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/UpgradeStatus'
...
    OkEnvelope:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          const: true
        data:
          type: object
          additionalProperties: true
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque